tester utils error handling
